limit asm to Linux
